Grouped and refactored functions and services into single files
