+++ server/__init__.py
@@ -2,8 +2,7 @@ from flask import Flask
 from flask_cors import CORS
 from .config.env_config import UI_URL
 from .endpoints.hello import hello_bp
-from .endpoints.user_input import user_input_bp
-from .endpoints.image_upload import image_upload_bp
+from .endpoints.chat import chat_bp
 from .endpoints.parcel_finder import parcel_finder_bp
 import shutil
 import os
@@ -17,8 +16,7 @@ def create_app():
 
     # Register Blueprints
     app.register_blueprint(hello_bp)
-    app.register_blueprint(user_input_bp)
-    app.register_blueprint(image_upload_bp)
+    app.register_blueprint(chat_bp)
     app.register_blueprint(parcel_finder_bp)
     
     return app
