+++ server/services/chat_service.py
@@ -1,8 +1,13 @@
 from PIL import Image
 from ..utils.llm_utils import load_prompt_from_json
-from ..utils.image_utils import save_image_and_get_path
+from ..utils.chat_utils import save_image_and_get_path
 from ..config.chat_init_config import CHAT as chat
 from ..config.constants import PROMPT_LIST_FILE
+from ..config.chat_init_config import CHAT as chat
+
+def generate_user_response(user_input: str) -> str:
+    response = chat.send_message(user_input,)
+    return response.text
 
 def get_image_description(file):
     """
@@ -14,3 +19,4 @@ def get_image_description(file):
     response = chat.send_message([image, load_prompt_from_json(PROMPT_LIST_FILE, is_image_desc_prompt=True)],)
 
     return response.text
+
